+++ app/src/main/java/com/example/team25/MyApplication.kt
@@ -1,18 +1,9 @@
 package com.example.team25
 
 import android.app.Application
-import com.example.team25.data.dao.HospitalDao
-import com.example.team25.domain.HospitalDomain
-import com.google.gson.Gson
-import com.google.gson.reflect.TypeToken
 import com.kakao.sdk.common.KakaoSdk
 import com.kakao.vectormap.KakaoMapSdk
 import dagger.hilt.android.HiltAndroidApp
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import java.io.InputStreamReader
-import javax.inject.Inject
 
 @HiltAndroidApp
 class MyApplication : Application() {
