+++ app/src/main/java/com/example/team25/domain/model/HospitalDomain.kt
@@ -1,4 +1,4 @@
-package com.example.team25.domain
+package com.example.team25.domain.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
