--- app/src/main/java/com/example/team25/ui/login/SocialLoginUiState.kt
@@ -1,11 +1,0 @@
-package com.example.team25.ui.login
-
-sealed interface SocialLoginUiState {
-    object SocialLoginUiSuccess : SocialLoginUiState
-
-    object SocialLoginUiFail : SocialLoginUiState
-
-    object IDle : SocialLoginUiState
-
-    object KakaoSocialLoginUi : SocialLoginUiState
-}
