+++ app/src/main/java/com/example/team25/ui/reservation/ReservationStep9Fragment.kt
@@ -8,7 +8,7 @@ import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.team25.R
 import com.example.team25.databinding.FragmentReservationStep9Binding
-import com.example.team25.domain.ManagerDomain
+import com.example.team25.domain.model.ManagerDomain
 import com.example.team25.ui.reservation.adapters.ManagerRecyclerViewAdapter
 import com.example.team25.ui.reservation.interfaces.OnManagerClickListener
