+++ app/src/main/java/com/example/team25/ui/reservation/interfaces/SearchHospitalService.kt
@@ -1,6 +1,6 @@
 package com.example.team25.ui.reservation.interfaces
 
-import com.example.team25.domain.HospitalDomain
+import com.example.team25.domain.model.HospitalDomain
 
 interface SearchHospitalService {
     suspend fun getSearchedResult(
